requirements.txt
Django==5.0.6
easyocr==1.7.1
fitz==0.0.1.dev2
matplotlib==3.8.4
numpy==1.26.4
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
Pillow==10.3.0
protobuf==4.25.3
pytesseract==0.3.10
Requests==2.32.2